% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatures.R
\name{getFeatures}
\alias{getFeatures}
\title{Parses a genome file in GFF format into a list of \code{GRangesList}s organized by gene and genomic feature.}
\usage{
getFeatures(
  gff,
  prefix = "KY2019:",
  tssflank = c(1107, 107),
  tsswindow = 10000,
  ttswindow = 10000,
  ...
)
}
\arguments{
\item{gff}{Path to a genome file in GFF format.}

\item{prefix}{Prefix to append to gene IDs.}

\item{tssflank}{Numeric vector of length 2 giving the window in bp upstream and downstream of the TSS to use as the promoter.}

\item{tsswindow}{How many bp upstream of the promoter to consider as upstream of a gene.}

\item{ttswindow}{How many bp downstream of the TTS to consider as downstream of a gene.}
}
\value{

}
\description{
Parses a genome file in GFF format into a list of \code{GRangesList}s organized by gene and genomic feature.
}
\seealso{
\code{\link{mergeMotifs}}, \code{\link{motifmatchr::matchMotifs}}
}
