% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatures.R
\name{getOverlapMat}
\alias{getOverlapMat}
\title{Converts gene-peak table to a logical matrix of associations}
\usage{
getOverlapMat(overlaps)
}
\arguments{
\item{overlaps}{A two-column matrix with peak IDs in the first column and gene IDs in the second column. It is intended to be used with the output of \code{getOverlaps}.}
}
\value{
A logical matrix of gene-peak associations with genes as columns and peaks as rows.
}
\description{
Converts gene-peak table to a logical matrix of associations
}
\seealso{
\code{\link{getOverlaps}}
}
