% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatures.R
\name{importPeaks}
\alias{importPeaks}
\title{Import peaks file with names.}
\usage{
importPeaks(con, prefix = "", ...)
}
\arguments{
\item{con}{The input file}

\item{prefix}{A string appended to the start of each name.}

\item{...}{Arguments passed to \code{\link{rtracklayer::import}}}
}
\description{
Import peaks file with names.
}
\seealso{
\code{\link{namePeaks}}, \code{\link{rtracklayer::import}}
}
