% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatures.R
\name{namePeaks}
\alias{namePeaks}
\title{Assigns names to peaks.}
\usage{
namePeaks(peaks, prefix = "")
}
\arguments{
\item{peaks}{A GRanges object}

\item{prefix}{A string appended to the start of each name.}
}
\description{
Assigns names to peaks.
}
